\name{internal-var}
\alias{ClusterVCV}
\alias{GmatInv}
\alias{GmatInvLM}
\alias{GmatInvRWE}
\alias{hmat}
\alias{hmatModel}
\alias{SmatRobust}
\alias{SmatCluster}
\alias{VarRWEcluster}
\alias{VarRWErobust}
\title{Internal functions for estimating variances}
\description{These are helper function to calculate various variance-covariance component matrices.}

\keyword{internal}
