\name{GSSU-package}
\alias{GSSU-package}
\alias{GSSU}
\docType{package}
\title{Estimate the sample-weighted effect as in Gibbons et al. (2018)}
\description{This package estimates the interaction-weighted and
  regression-weighted effects (IWE and RWE respectively) as derived in
  Gibbons, Suarez Serrato, and Urbancic (2018). It also implements specification tests for both estimators and the Wald and score tests for heterogeneous treatment effects.}
\details{
\tabular{ll}{
Package: \tab GSSU
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-06-28\cr
License: \tab GPL-3\cr
}
The key functions are \code{EstimateIWE}, \code{EstimateRWE}, \code{SpecTest}, \code{WaldTestIWE}, and \code{ScoreTest}.
}
\author{
Charles E. Gibbons
Maintainer: Charlie Gibbons <charlie.gibbons@brattle.com>
}
\references{
Gibbons, Charles E., Juan Carlos Suarez Serrato, and Michael
B. Urbancic. 2018. "Broken or Fixed Effects?" \emph{Journal of
	Econometric Methods}.}
\keyword{ package }
